feat: Work around for flutter 3.22.3 -> 3.24.2
There was a big internal migration for RenderParagraph to implement paragraph selection that breaks this. Trying to use the copy over the changes and re-implementing the custom selection logic breaks functionality in the obelisk application. A poc failed, see branch `flutter_3_24_2`. This seems to work for now, but may need to revert and switch to the old version of flutter.